+++ ricaun.AppBundleTool/AppBundle/AppBundleAccess.cs
@@ -1,0 +1,8 @@
+﻿namespace ricaun.AppBundleTool.AppBundle
+{
+    public enum AppBundleAccess
+    {
+        Allow,      // Access is permitted
+        Admin       // Access requires administrator privileges
+    }
+}